Add logger link
