+++ packages/client/src/core/create-client.ts
@@ -1,92 +1,4 @@
-import { WebSocket } from "ws";
-
-type LinkContext = {
-    path: string;
-    args: any;
-    next: () => Promise<any>;
-};
-type Link = (context: LinkContext) => Promise<any>;
-
-type TransportLinkConfiguration = {
-    baseUrl?: string;
-};
-
-export const httpLink =
-    (config?: TransportLinkConfiguration): Link =>
-    async ({ path, args, next }) => {
-        return fetch((config ? config.baseUrl ?? "" : "") + path, {
-            method: "POST",
-            body: JSON.stringify(args),
-            headers: { "Content-Type": "application/json" },
-            cache: "no-cache",
-            credentials: "include",
-        })
-            .then((res) => res.json())
-            .catch((err) => err);
-    };
-
-const waitForOpenConnection = (ws: WebSocket): Promise<void> => {
-    return new Promise((resolve, reject) => {
-        if (ws.readyState === WebSocket.OPEN) {
-            resolve();
-        } else {
-            const onOpen = () => {
-                ws.removeEventListener("open", onOpen);
-                resolve();
-            };
-            ws.addEventListener("open", onOpen);
-        }
-    });
-};
-
-export const wsLink = (config?: TransportLinkConfiguration): Link => {
-    const ws = new WebSocket(config?.baseUrl ?? "");
-    return async ({ path, args, next }) => {
-        await waitForOpenConnection(ws);
-        const message = {
-            path,
-            method: "POST",
-            input: args,
-        };
-        return new Promise((resolve, reject) => {
-            const listener = (event: any) => {
-                const response = JSON.parse(event.data);
-                if (response.status >= 200 && response.status < 300) {
-                    resolve(response.data);
-                } else {
-                    reject(response);
-                }
-                ws.removeEventListener("message", listener);
-            };
-            ws.addEventListener("message", listener);
-            ws.send(JSON.stringify(message));
-        });
-    };
-};
-
-export const loggerLink =
-    (): Link =>
-    async ({ path, args, next }) => {
-        console.log("%c\n=== Request ===", "color: blue; font-weight: bold; font-size: 16px;");
-        console.log("%cPath: %c" + path, "color: cyan; font-weight: bold;", "color: green; font-weight: bold;");
-        console.log(
-            "%cInput: %c" + JSON.stringify(args, null, 2),
-            "color: cyan; font-weight: bold;",
-            "color: yellow; font-weight: bold;",
-        );
-
-        const response = await next();
-
-        console.log("%c\n=== Response ===", "color: blue; font-weight: bold; font-size: 16px;");
-        console.log("%cPath: %c" + path, "color: cyan; font-weight: bold;", "color: green; font-weight: bold;");
-        console.log(
-            "%cData: %c" + JSON.stringify(response, null, 2),
-            "color: cyan; font-weight: bold;",
-            "color: yellow; font-weight: bold;",
-        );
-
-        return response;
-    };
+import { Link } from "./links";
 
 export type CreateClientConfiguration = {
     links: Link[];
