+++ packages/client/src/core/index.ts
@@ -1,1 +1,2 @@
 export * from "./create-client";
+export * from "./links";
